+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 3.8.3
+- Add scDART
+- Fix bug for non-aligned datasets in `commando` module
+- Various visualizations in `evaluation` module
+
 # 3.8.2
 - Fixed a bug concerning min-max normalization in the `compute_distances` function
 - Implement SHAP and add visualizations